+++ ui-components/src/main/kotlin/team/duckie/quackquack/ui/component/image.kt
@@ -42,7 +42,6 @@ import team.duckie.quackquack.ui.color.QuackColor
 import team.duckie.quackquack.ui.color.animateQuackColorAsState
 import team.duckie.quackquack.ui.component.QuackSelectableImageType.CheckOverlay
 import team.duckie.quackquack.ui.component.QuackSelectableImageType.TopEndCheckBox
-import team.duckie.quackquack.ui.component.internal.QuackSurface
 import team.duckie.quackquack.ui.icon.QuackIcon
 import team.duckie.quackquack.ui.modifier.quackClickable
 import team.duckie.quackquack.ui.util.DpSize
